cloud_run/requirements.txt
# TechnoBrain-MENDAN Cloud Run Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Google Cloud APIs
google-cloud-speech==2.24.0
google-cloud-storage==2.14.0
google-cloud-secret-manager==2.18.0

# Google Sheets
gspread==6.0.0
google-auth==2.27.0

# HTTP Client
httpx==0.26.0

# Data Validation
pydantic==2.6.0

# Anthropic Claude API
anthropic==0.40.0
